Fix server Crash
