+++ projects/pswg/src/main/java/com/parzivail/pswg/entity/FragmentationGrenadeEntity.java
@@ -3,31 +3,20 @@ package com.parzivail.pswg.entity;
 import com.parzivail.pswg.Resources;
 import com.parzivail.pswg.client.sound.SoundHelper;
 import com.parzivail.pswg.container.*;
-import com.parzivail.pswg.item.ExplosionSoundGroup;
 import com.parzivail.util.data.PacketByteBufHelper;
 import io.netty.buffer.Unpooled;
-import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
-import net.fabricmc.fabric.api.networking.v1.S2CPlayChannelEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.block.BlockState;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.network.ClientPlayNetworkHandler;
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.thrown.ThrownEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.PacketByteBuf;
-import net.minecraft.network.packet.s2c.play.EntityAttributesS2CPacket;
-import net.minecraft.network.packet.s2c.play.EntityS2CPacket;
-import net.minecraft.particle.ParticleTypes;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
-import net.minecraft.sound.SoundEvent;
-import net.minecraft.text.Text;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.hit.BlockHitResult;
@@ -41,7 +30,6 @@ import net.minecraft.world.World;
 import net.minecraft.world.explosion.Explosion;
 
 import java.util.List;
-import java.util.concurrent.ThreadLocalRandom;
 
 public class FragmentationGrenadeEntity extends ThrowableExplosive
 {
@@ -137,42 +125,7 @@ public class FragmentationGrenadeEntity extends ThrowableExplosive
 		}
 	}
 
-	public static void createSparkParticles(World world, Vec3d pos, Entity entity, boolean yCollision)
-	{
-		for (int i = 0; i < Random.create().nextBetween(50, 80); i++)
-		{
-			double vx = world.random.nextGaussian() * 0.5;
-			double vz = world.random.nextGaussian() * 0.5;
-			double vy;
 
-			if (yCollision)
-				vy = Math.abs(world.random.nextGaussian() * 0.8);
-			else
-				vy = world.random.nextGaussian() * 0.4;
-			world.addParticle(SwgParticleTypes.FRAGMENTATION_GRENADE_SPARK, pos.x, pos.y, pos.z, vx, vy, vz);
-		}
-	}
-
-	public static void handleExplosion(MinecraftClient client, ClientPlayNetworkHandler handler, PacketByteBuf buf, PacketSender responseSender)
-	{
-		boolean isStart = buf.readBoolean();
-		if (isStart)
-		{
-			var entityId = buf.readInt();
-			var world = client.world;
-			var entity = world.getEntityById(entityId);
-			if (entity instanceof FragmentationGrenadeEntity fragmentationGrenade)
-				fragmentationGrenade.SHOULD_RENDER = false;
-		}
-		else
-		{
-			var world = client.world;
-			var pos = PacketByteBufHelper.readVec3d(buf);
-			var entityId = buf.readInt();
-			var yCollision = buf.readBoolean();
-			client.execute(() -> createSparkParticles(world, pos, world.getEntityById(entityId), yCollision));
-		}
-	}
 
 	@Override
 	public void tick()
